requirements.txt
# File: requirements.txt (Stateless Edition)
Flask==2.3.2
gunicorn==20.1.0
gevent==24.2.1
python-dotenv==1.0.0
gspread==5.8.0
google-auth-oauthlib==1.0.0
requests==2.31.0